Implement optimized wildcard request dispatching
This reduces the sub/unsub churn caused by a naive implementation of the
request method. This is the method used in the official nats.go client.

When a timeout on request happens, the request method is responsible for
cleaning up. The request handler will always remove the subscription
handle when a message has been received.