Add request_with_timeout method
This will contain all of the log for the request operation. The existing
request method will call request_with_timeout using a u64::MAX seconds
timeout to simulate an infinite timeout (or never timeout).

This also adds a From impl for the tokio time elapsed error. This feels
a bit strange and could certainly be reworked here.