Wrap logic in a request type that's droppable
This ensures any success or failure in the request process will remove
the reply_to from the request_inbox_mapping.

I don't like the blocking Drop implementation, however, it is only
blocking for a mutex which seems ok, but not great. There's also room to
optimize since the success path will be auto-removed from the mapping.